Hello World in Java;
The main class is static, meaning it doesn't need to be instantiated. It is a method acting as an entry point for the program. The println method prints a line with a new line character to the console. System is an existing class, which has a stream called out. This is used to access the println method.