Operators
Arithmetic operators are basically used in the same way as in other languages. They can be used in combinations and obey the rules of maths. Brackets can also be part of calculations. Notice the use of 9.0 instead of 9 in the division. This is in order to forego an issue with the truncation of integers in mathematical operations. The result of effect of this can be seen in the last line where it is intentionally provoked. A division which would result in a floating point number, is converted into an integer by simply ignoring digits after the period. A similar thing would happen in the division of 9 / 2. The result of the division would be 4, because java doesn't return the digits that would be after the period.

Increment and decrement can be used to quickly increase or decrease the value of a numeric variable by one. Prefix (operator in front of variable) manipulates the value and then uses it in the expression whereas postfix (variable in front of operator) uses the variable in the expression first and then manipulates the value.

Assignment operators can also be used to quickly maniupulate variables. The operators shown here can be used to quickly assign the result of the calculation to the variable. Say z = 5; z *= 3; z is now equal to 15.