Conditional Statements
One of the most used conditional statements used is the if statement. The structure can be seen from lines 55 - 57. In case the expression inside the if-statement is on a single line, there is no need to use curly brackets. The comparison operators used in java are similar to most other high-level programming languages. The else keyword does also exist. if - statements can be nested. The else if operator also exists. In order to combine multiple conditions inside a singe check one can use the or ||, and &&, and the not ! operator. Switch statements are a way to quickly check a row of values for a given condition. The switch - expression can also be used to return a value. Multiple values can be checked in a single case.